refactor(NODE-3717): test reorganization part 6 (#3090)
